location has been moved out from modules
