+++ catch/libcerberus/text/text_iterator_modules/location.cpp
@@ -1,5 +1,5 @@
 #include <cerberus/debug/debug.hpp>
-#include <cerberus/text/text_iterator_modules/location.hpp>
+#include <cerberus/text/location.hpp>
 
 using namespace cerb::debug;
 using Location = cerb::text::Location<>;
